# api/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Fundly LOS API
  version: 1.0.0-p0
  summary: 'Loan origination system for Nigerian commercial banks: Phase 0 foundation.'
  description: 'Design-first contract for /api/v1 (TRD §10). Conventions:


    - **Auth**: staff use a same-origin Sanctum session cookie (+ XSRF token); service and partner principals use bearer tokens whose abilities
    can only narrow their role assignments.

    - **Idempotency-Key** is required on every POST that creates or causes an effect (`x-idempotency-key: required`). A replay returns the stored
    response (header `Idempotent-Replayed: true`); the same key with a different body returns 409 `idempotency-key-reused`.

    - **Concurrency**: resources carry `ETag`; PATCH requires `If-Match` (428 if missing, 412 if stale).

    - **Errors**: RFC 9457 `application/problem+json` with `type` (URN), `code`, `correlation_id`, and `errors` for validation failures.

    - **Lists**: cursor pagination with `page[size]` (≤ 100) and `page[after]`; scope filtering is always applied.

    - **Tracing**: `X-Correlation-Id` is accepted or generated and always echoed.

    - **Maker-checker** actions answer 202 with the change request; nothing changes until a different user approves it.

    - `x-permission` names the permission checked at the route; `x-step-up` marks operations needing a re-authentication within 5 minutes.

    '
  license:
    name: Proprietary (Atheris Limited)
    identifier: LicenseRef-Atheris-Proprietary
servers:
...
paths:
...
  /api/v1/applications:
    get:
      summary: 'List applications in scope'
      operationId: listApplications
      tags:
        - Applications
      x-permission: 'application:view'
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageAfter'
        - name: 'filter[status]'
          in: query
          required: false
          schema:
            type: string
          description: 'Comma-separated canonical statuses'
        - name: 'filter[open]'
          in: query
          required: false
          schema:
            type: string
          description: 'true: exclude terminal'
        - name: 'filter[mine]'
          in: query
          required: false
          schema:
            type: string
          description: 'true: originated by me'
        - name: 'filter[product_key]'
          in: query
          required: false
          schema:
            type: string
          description: 'Product key'
        - name: 'filter[party_id]'
          in: query
          required: false
          schema:
            type: string
          description: 'Applications with this party'
        - name: 'filter[q]'
          in: query
          required: false
          schema:
            type: string
          description: 'Reference or applicant name'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApplicationSummary'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                required:
                  - data
                  - meta
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
    post:
      summary: 'Create a draft application'
      operationId: createApplication
      tags:
        - Applications
      x-permission: 'application:originate'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreate'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - data
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
      x-idempotency-key: required
...
  '/api/v1/applications/{id}':
    get:
      summary: 'Get an application'
      operationId: getApplication
      tags:
        - Applications
      x-permission: 'application:view'
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - data
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
    patch:
      summary: 'Amend captured fields before approval (field history + provenance)'
      operationId: amendApplication
      tags:
        - Applications
      x-permission: 'application:originate'
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requested_amount:
                  type:
                    - string
                    - 'null'
                tenor_months:
                  type:
                    - integer
                    - 'null'
                purpose:
                  type:
                    - string
                    - 'null'
                repayment_frequency:
                  type:
                    - string
                    - 'null'
                data:
                  type: object
                source:
                  type: string
                  enum:
                    - cba
                    - extraction
                    - self_service
                    - staff
                    - partner
                  default: staff
                source_ref:
                  type:
                    - string
                    - 'null'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - data
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
  '/api/v1/applications/{id}/applicants':
    post:
      summary: 'Add a joint applicant, guarantor or co-signer'
      operationId: addApplicant
      tags:
        - Applications
      x-permission: 'application:originate'
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                party_id:
                  type: string
                  format: uuid
                role:
                  type: string
                  enum:
                    - joint
                    - guarantor
                    - co_signer
              required:
                - party_id
                - role
      responses:
        201:
          description: Added
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - data
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
      x-idempotency-key: required
  '/api/v1/applications/{id}/applicants/{partyId}':
    delete:
      summary: 'Remove a non-primary applicant'
      operationId: removeApplicant
      tags:
        - Applications
      x-permission: 'application:originate'
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: partyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        200:
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - data
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
  '/api/v1/applications/{id}/actions/{action}':
    post:
      summary: 'Lifecycle action (submit, recommend, hold, resume, return, resubmit, withdraw, cancel)'
      operationId: actOnApplication
      tags:
        - Applications
      x-permission: 'application:view'
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum:
              - submit
              - withdraw
              - cancel
              - hold
              - resume
              - return
              - resubmit
              - recommend
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason_code:
                  type:
                    - string
                    - 'null'
                  pattern: '^[A-Z0-9_]{2,64}$'
                reason_text:
                  type:
                    - string
                    - 'null'
                return_to:
                  type:
                    - string
                    - 'null'
      responses:
        200:
          description: Transitioned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - data
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
      x-idempotency-key: required
      description: 'The command checks the action-specific permission: application:originate (submit, hold, resume, resubmit, withdraw, cancel) or application:recommend (recommend, return). withdraw, cancel, hold and return need reason_code.'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-fundly_session
      description: Sanctum stateful SPA session (HttpOnly, Secure, SameSite=Strict) + X-XSRF-TOKEN header on unsafe methods.
    bearerToken:
      type: http
      scheme: bearer
      description: Sanctum personal access token for service and partner accounts.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        pattern: ^[A-Za-z0-9._:-]{8,255}$
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
      description: ETag from the last read
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        pattern: ^[A-Za-z0-9._:-]{8,64}$
    PageSize:
      name: page[size]
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    PageAfter:
      name: page[after]
      in: query
      schema:
        type: string
      description: Opaque cursor from meta.page.next_cursor
  headers:
    ETag:
      schema:
        type: string
      description: Entity tag for If-Match
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        correlation_id:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required:
      - type
      - title
      - status
      - detail
      - code
      - correlation_id
      description: RFC 9457 problem details. `type` is a URN from the problem catalogue.
    PageMeta:
      type: object
      properties:
        page:
          type: object
          properties:
            size:
              type: integer
            next_cursor:
              type:
              - string
              - 'null'
            has_more:
              type: boolean
          required:
          - size
          - next_cursor
          - has_more
      required:
      - page
    Money:
      type: object
      properties:
        amount:
          type: string
          pattern: ^-?[0-9]+\.[0-9]{4}$
        currency:
          type: string
          pattern: ^[A-Z]{3}$
      required:
      - amount
      - currency
      description: Decimal string at scale 4 + ISO 4217 code. Never a float.
...
    ApplicationCreate:
      type: object
      properties:
        legal_entity_id:
          type: string
          format: uuid
        org_unit_id:
          type: string
          format: uuid
        product_key:
          type: string
        primary_party_id:
          type: string
          format: uuid
        channel:
          type: string
          enum:
            - staff
            - api
          default: staff
        requested_amount:
          type:
            - string
            - 'null'
        tenor_months:
          type:
            - integer
            - 'null'
        purpose:
          type:
            - string
            - 'null'
        repayment_frequency:
          type:
            - string
            - 'null'
        data:
          type: object
          description: 'Product-specific captured fields (flat, scalar values)'
      required:
        - legal_entity_id
        - org_unit_id
        - product_key
        - primary_party_id
    ApplicationSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        status:
          type: string
          enum:
            - draft
            - submitted
            - pre_qualified
            - kyc_screening
            - documentation
            - assessment
            - recommended
            - approval
            - approved
            - counter_offered
            - offer_issued
            - accepted
            - conditions_precedent
            - ready_for_disbursement
            - disbursing
            - booked
            - declined
            - expired
            - on_hold
            - returned_for_rework
            - withdrawn
            - cancelled
        product:
          type: object
          properties:
            id:
              type: string
              format: uuid
            key:
              type: string
            name:
              type: string
            version_id:
              type: string
              format: uuid
          required:
            - id
            - key
            - name
            - version_id
        segment:
          type: string
        channel:
          type: string
        primary_applicant:
          type: object
          properties:
            party_id:
              type: string
              format: uuid
            display_name:
              type: string
          required:
            - party_id
            - display_name
        requested_amount:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        currency:
          type: string
        tenor_months:
          type:
            - integer
            - 'null'
        legal_entity_id:
          type: string
          format: uuid
        org_unit_id:
          type: string
          format: uuid
        originator_id:
          type: string
          format: uuid
        submitted_at:
          type:
            - string
            - 'null'
          format: date-time
        status_changed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        version:
          type: integer
      required:
        - id
        - reference
        - status
        - product
        - segment
        - channel
        - primary_applicant
        - requested_amount
        - currency
        - tenor_months
        - legal_entity_id
        - org_unit_id
        - originator_id
        - submitted_at
        - status_changed_at
        - created_at
        - version
    Application:
      allOf:
        - $ref: '#/components/schemas/ApplicationSummary'
        - type: object
          properties:
            purpose:
              type:
                - string
                - 'null'
            repayment_frequency:
              type:
                - string
                - 'null'
            data:
              type: object
            resume_to:
              type:
                - string
                - 'null'
              enum:
                - draft
                - submitted
                - pre_qualified
                - kyc_screening
                - documentation
                - assessment
                - recommended
                - approval
                - approved
                - counter_offered
                - offer_issued
                - accepted
                - conditions_precedent
                - ready_for_disbursement
                - disbursing
                - booked
                - declined
                - expired
                - on_hold
                - returned_for_rework
                - withdrawn
                - cancelled
                - null
            return_to:
              type:
                - string
                - 'null'
              enum:
                - draft
                - submitted
                - pre_qualified
                - kyc_screening
                - documentation
                - assessment
                - recommended
                - approval
                - approved
                - counter_offered
                - offer_issued
                - accepted
                - conditions_precedent
                - ready_for_disbursement
                - disbursing
                - booked
                - declined
                - expired
                - on_hold
                - returned_for_rework
                - withdrawn
                - cancelled
                - null
            closed_at:
              type:
                - string
                - 'null'
              format: date-time
            close_reason_code:
              type:
                - string
                - 'null'
            expires_at:
              type:
                - string
                - 'null'
              format: date-time
            applicants:
              type: array
              items:
                type: object
                properties:
                  party_id:
                    type: string
                    format: uuid
                  role:
                    type: string
                    enum:
                      - primary
                      - joint
                      - guarantor
                      - co_signer
                  party_type:
                    type: string
                  display_name:
                    type: string
                required:
                  - party_id
                  - role
                  - party_type
                  - display_name
            completeness:
              type: object
              properties:
                percent:
                  type: integer
                  minimum: 0
                  maximum: 100
                missing:
                  type: array
                  items:
                    type: string
                checklist:
                  type: array
                  items:
                    type: object
                    properties:
                      code:
                        type: string
                      name:
                        type: string
                      mandatory:
                        type: boolean
                    required:
                      - code
                      - name
                      - mandatory
              required:
                - percent
                - missing
                - checklist
          required:
            - purpose
            - data
            - applicants
            - completeness
...